test(e2e): stabilize tests — auth wait, snackbar fix, pickDateTime consolidation
- Add waitFor("Aanstaande trainingen") after page.goto() in beforeEach of
  crud-matches, crud-training, login, and previous-events specs to prevent
  timeout failures when backend API retries are still in progress
- Fix deleteTraining snackbar assertion in training-utils.ts to use
  .filter({ hasText: ... }) pattern (matches deleteMatch in match-utils.ts),
  preventing strict-mode violations with multiple alerts
- Extract pickDateTime into utils.ts as shared export; remove duplicate
  implementations from training-utils.ts and match-utils.ts
