+++ e2e/src/playwright/tests/crud-training.spec.ts
@@ -11,6 +11,7 @@ test.describe("Training ", () => {
   test.beforeEach(async ({ page }) => {
     // Go to the starting url before each test.
     await page.goto(HOST);
+    await page.getByText("Aanstaande trainingen").waitFor({ state: "visible" });
   });
 
   test("CRUD normal training", async ({ page }) => {
